Sort on sales package id
