--- src/Service.Host/client/src/actions/salesPackage.js
@@ -1,9 +1,0 @@
-import { salesPackageActionTypes as actionTypes } from './index';
-import UpdateApiActions from './UpdateApiActions';
-import * as itemTypes from '../itemTypes';
-
-export default new UpdateApiActions(
-    itemTypes.salesPackage.actionType,
-    itemTypes.salesPackage.uri,
-    actionTypes
-);
