--- src/Service.Host/client/src/actions/salesPackages.js
@@ -1,9 +1,0 @@
-import { salesPackagesActionTypes as actionTypes } from './index';
-import FetchApiActions from './FetchApiActions';
-import * as itemTypes from '../itemTypes';
-
-export default new FetchApiActions(
-    itemTypes.salesPackages.actionType,
-    itemTypes.salesPackages.uri,
-    actionTypes
-);
