+++ src/Service.Host/client/src/actions/salesPackages/salesPackage.js
@@ -1,0 +1,9 @@
+import { salesPackageActionTypes as actionTypes } from '../index';
+import UpdateApiActions from '../UpdateApiActions';
+import * as itemTypes from '../../itemTypes';
+
+export default new UpdateApiActions(
+    itemTypes.salesPackage.actionType,
+    itemTypes.salesPackage.uri,
+    actionTypes
+);
