+++ src/Service.Host/client/src/components/sernosSequences/SernosSequences.js
@@ -1,21 +1,21 @@
 import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { Loading, EntityList, CreateButton } from '@linn-it/linn-form-components-library';
-import Page from '../containers/Page';
-import { sortEntityList } from '../helpers/utilities';
+import Page from '../../containers/Page';
+import { sortEntityList } from '../../helpers/utilities';
 
-function VatCodes({ vatCodes, loading }) {
+function SernosSequences({ sernosSequences, loading }) {
     return (
         <Page>
             {loading ? (
                 <Loading />
             ) : (
                 <Fragment>
-                    <CreateButton createUrl="/products/maint/vat-codes/create" />
+                    <CreateButton createUrl="/products/maint/sernos-sequences/create" />
                     <EntityList
-                        title="Vat Codes"
-                        entityList={sortEntityList(vatCodes, 'code')}
-                        entityId="code"
+                        title="Sernos Sequences"
+                        entityList={sortEntityList(sernosSequences, 'sequenceName')}
+                        entityId="sequenceName"
                         loading={loading}
                         descriptionFieldName="description"
                     />
@@ -25,9 +25,9 @@ function VatCodes({ vatCodes, loading }) {
     );
 }
 
-VatCodes.propTypes = {
-    vatCodes: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+SernosSequences.propTypes = {
+    sernosSequences: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
     loading: PropTypes.bool.isRequired
 };
 
-export default VatCodes;
+export default SernosSequences;
