+++ src/Service.Host/client/src/containers/salesPackages/SalesPackage.js
@@ -1,7 +1,7 @@
 ﻿import { connect } from 'react-redux';
 import getSingleErrorMessage from '../../selectors/fetchErrorSelectors';
 import SalesPackage from '../../components/salesPackages/SalesPackage';
-import salesPackagesActions from '../../actions/salesPackage';
+import salesPackagesActions from '../../actions/salesPackages/salesPackage';
 import initialiseOnMount from '../common/initialiseOnMount';
 import salesPackageSelectors from '../../selectors/salesPackageSelectors';
