+++ src/Service.Host/client/src/containers/typesOfSale/CreateTypeOfSale.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import initialiseOnMount from '../common/initialiseOnMount';
 import typeOfSaleActions from '../../actions/typeOfSaleActions';
-import TypeOfSale from '../../components/TypeOfSale';
+import TypeOfSale from '../../components/typesOfSale/TypeOfSale';
 import getSingleErrorMessage from '../../selectors/fetchErrorSelectors';
 import typeOfSaleSelectors from '../../selectors/typeOfSaleSelectors';
