+++ src/Service.Host/index.html
@@ -1,19 +1,19 @@
-﻿<!doctype html>
+﻿<!DOCTYPE html>
 <html>
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Products (local hotload)</title>
-</head>
-<body>
-    <div id="root"></div>
-    <script type="text/javascript">
-        window.APPLICATION_SETTINGS = {
-            authorityUri: 'https://www-sys.linn.co.uk/auth/',
-            appRoot: 'http://localhost:51798',
-            proxyRoot: 'http://app.linn.co.uk'
-        };
-    </script>
-    <script src="/products/maint/build/app.js"></script>
-</body>
+    <head>
+        <meta charset="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <title>Products (local hotload)</title>
+    </head>
+    <body>
+        <div id="root"></div>
+        <script type="text/javascript">
+            window.APPLICATION_SETTINGS = {
+                authorityUri: 'https://www-sys.linn.co.uk/auth/',
+                appRoot: 'http://localhost:5050',
+                proxyRoot: 'http://app.linn.co.uk'
+            };
+        </script>
+        <script src="/products/maint/build/app.js"></script>
+    </body>
 </html>
