+++ tests/Integration/Service.Tests/Service.Tests.csproj
@@ -9,7 +9,7 @@
   <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="4.19.4" />
     <PackageReference Include="FluentValidation" Version="8.2.0" />
-    <PackageReference Include="Linn.Common.Facade" Version="4.0.0" />
+    <PackageReference Include="Linn.Common.Facade" Version="5.0.0" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
     <PackageReference Include="MSTest.TestAdapter" Version="1.2.0" />
     <PackageReference Include="MSTest.TestFramework" Version="1.2.0" />
