+++ tests/Unit/Facade.Tests/Facade.Tests.csproj
@@ -8,7 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="FluentAssertions" Version="4.19.4" />
-    <PackageReference Include="Linn.Common.Facade" Version="4.0.0" />
+    <PackageReference Include="Linn.Common.Facade" Version="5.0.0" />
     <PackageReference Include="Linn.Common.Persistence" Version="2.0.1" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
     <PackageReference Include="MSTest.TestAdapter" Version="1.2.0" />
